重构 initDB 和 createTables
