+++ main.go
@@ -3,6 +3,7 @@ package main
 import (
 	"database/sql"
 	"fmt"
+	"goblog/pkg/database"
 	"goblog/pkg/logger"
 	"goblog/pkg/route"
 	"goblog/pkg/types"
@@ -11,42 +12,14 @@ import (
 	"strconv"
 	"strings"
 	"text/template"
-	"time"
 	"unicode/utf8"
 
-	"github.com/go-sql-driver/mysql"
 	"github.com/gorilla/mux"
 )
 
 var router *mux.Router
 
 var db *sql.DB
-
-func initDB() {
-	var err error
-	config := mysql.Config{
-		User:                 "root",
-		Passwd:               "qwer1234",
-		Addr:                 "127.0.0.1:3306",
-		Net:                  "tcp",
-		DBName:               "goblog",
-		AllowNativePasswords: true,
-	}
-	//准备数据库连接池
-	db, err = sql.Open("mysql", config.FormatDSN())
-	logger.LogError(err)
-	//设置最大连接数
-	db.SetMaxOpenConns(25)
-	//设置最大空闲连接数
-	db.SetMaxIdleConns(25)
-	//设置每个连接的过期时间
-	db.SetConnMaxLifetime(5 * time.Minute)
-
-	//尝试连接，失败会报错
-	err = db.Ping()
-	logger.LogError(err)
-
-}
 
 func homeHandler(w http.ResponseWriter, r *http.Request) {
 	fmt.Fprint(w, "<h1>Hello,欢迎来到happyBlog</h1>")
@@ -502,22 +475,9 @@ func articlesCreateHandler(w http.ResponseWriter, r *http.Request) {
 	}
 }
 
-// 创建articles表
-func createTables() {
-	createArticlesSQL := `
-		CREATE TABLE IF NOT EXISTS articles(
-			id bigint(20) PRIMARY KEY AUTO_INCREMENT NOT NULL,
-			title varchar(255) COLLATE utf8mb4_unicode_ci NOT NULL,
-			body longtext COLLATE utf8mb4_unicode_ci
-		);
-	`
-	_, err := db.Exec(createArticlesSQL)
-	logger.LogError(err)
-}
-
 func main() {
-	initDB()       //初始化数据库
-	createTables() //创建表articles
+	database.Initialize()
+	db = database.DB
 
 	route.Initialize()
 	router = route.Router
